+++ tests/test_on_data/test_gin_behavior.py
@@ -1,12 +1,130 @@
 import unittest
+from pathlib import Path
 from datetime import datetime
-from parameterized import parameterized, param
 
+from parameterized import parameterized, param
 from pynwb import NWBHDF5IO
 from neuroconv import NWBConverter
-from neuroconv.datainterfaces import MovieInterface, DeepLabCutInterface
+from neuroconv.datainterfaces import MovieInterface, DeepLabCutInterface, SLEAPInterface
 
 from .setup_paths import OUTPUT_PATH, BEHAVIOR_DATA_PATH
+import sleap_io
+
+
+class TestSLEAPInterface(unittest.TestCase):
+
+    savedir = OUTPUT_PATH
+
+    @parameterized.expand(
+        [
+            param(
+                data_interface=SLEAPInterface,
+                interface_kwargs=dict(
+                    file_path=str(BEHAVIOR_DATA_PATH / "sleap" / "predictions_1.2.7_provenance_and_tracking.slp")
+                ),
+            )
+        ]
+    )
+    def test_sleap_to_nwb_interface(self, data_interface, interface_kwargs):
+        nwbfile_path = str(self.savedir / f"{data_interface.__name__}.nwb")
+
+        interface = SLEAPInterface(**interface_kwargs)
+        metadata = interface.get_metadata()
+        metadata["NWBFile"].update(session_start_time=datetime.now().astimezone())
+        interface.run_conversion(nwbfile_path=nwbfile_path, overwrite=True, metadata=metadata)
+
+        slp_predictions_path = interface_kwargs["file_path"]
+        labels = sleap_io.load_slp(slp_predictions_path)
+
+        with NWBHDF5IO(path=nwbfile_path, mode="r", load_namespaces=True) as io:
+            nwbfile = io.read()
+            # Test matching number of processing modules
+            number_of_videos = len(labels.videos)
+            assert len(nwbfile.processing) == number_of_videos
+
+            # Test processing module naming as video
+            processing_module_name = "SLEAP_VIDEO_000_20190128_113421"
+            assert processing_module_name in nwbfile.processing
+
+            # For this case we have as many containers as tracks
+            # Each track usually represents a subject
+            processing_module = nwbfile.processing[processing_module_name]
+            processing_module_interfaces = processing_module.data_interfaces
+            assert len(processing_module_interfaces) == len(labels.tracks)
+
+            # Test name of PoseEstimation containers
+            extracted_container_names = processing_module_interfaces.keys()
+            for track in labels.tracks:
+                expected_track_name = f"track={track.name}"
+                assert expected_track_name in extracted_container_names
+
+            # Test one PoseEstimation container
+            container_name = f"track={track.name}"
+            pose_estimation_container = processing_module_interfaces[container_name]
+            # Test that the skeleton nodes are store as nodes in containers
+            expected_node_names = [node.name for node in labels.skeletons[0]]
+            assert expected_node_names == list(pose_estimation_container.nodes[:])
+
+            # Test that each PoseEstimationSeries is named as a node
+            for node_name in pose_estimation_container.nodes[:]:
+                assert node_name in pose_estimation_container.pose_estimation_series
+
+    @parameterized.expand(
+        [
+            param(
+                data_interface=SLEAPInterface,
+                interface_kwargs=dict(
+                    file_path=str(BEHAVIOR_DATA_PATH / "sleap" / "predictions_1.2.7_provenance_and_tracking.slp")
+                ),
+            )
+        ]
+    )
+    def test_sleap_to_nwb_converter(self, data_interface, interface_kwargs):
+        nwbfile_path = str(self.savedir / f"{data_interface.__name__}converter.nwb")
+
+        class TestConverter(NWBConverter):
+            data_interface_classes = dict(TestBehavior=data_interface)
+
+        converter = TestConverter(source_data=dict(TestBehavior=dict(interface_kwargs)))
+        metadata = converter.get_metadata()
+        metadata["NWBFile"].update(session_start_time=datetime.now().astimezone())
+        converter.run_conversion(nwbfile_path=nwbfile_path, overwrite=True, metadata=metadata)
+
+        slp_predictions_path = interface_kwargs["file_path"]
+        labels = sleap_io.load_slp(slp_predictions_path)
+
+        with NWBHDF5IO(path=nwbfile_path, mode="r", load_namespaces=True) as io:
+            nwbfile = io.read()
+            # Test matching number of processing modules
+            number_of_videos = len(labels.videos)
+            assert len(nwbfile.processing) == number_of_videos
+
+            # Test processing module naming as video
+            processing_module_name = "SLEAP_VIDEO_000_20190128_113421"
+            assert processing_module_name in nwbfile.processing
+
+            # For this case we have as many containers as tracks
+            # Each track usually represents a subject
+            processing_module = nwbfile.processing[processing_module_name]
+            processing_module_interfaces = processing_module.data_interfaces
+            assert len(processing_module_interfaces) == len(labels.tracks)
+
+            # Test name of PoseEstimation containers
+            extracted_container_names = processing_module_interfaces.keys()
+            for track in labels.tracks:
+                expected_track_name = f"track={track.name}"
+                assert expected_track_name in extracted_container_names
+
+            # Test one PoseEstimation container
+            container_name = f"track={track.name}"
+            pose_estimation_container = processing_module_interfaces[container_name]
+            # Test that the skeleton nodes are store as nodes in containers
+            expected_node_names = [node.name for node in labels.skeletons[0]]
+            assert expected_node_names == list(pose_estimation_container.nodes[:])
+
+            # Test that each PoseEstimationSeries is named as a node
+            for node_name in pose_estimation_container.nodes[:]:
+                assert node_name in pose_estimation_container.pose_estimation_series
 
 
 class TestDeepLabCutInterface(unittest.TestCase):
@@ -24,7 +142,7 @@ class TestDeepLabCutInterface(unittest.TestCase):
             )
         ]
     )
-    def test_convert_behaviordata_to_nwb(self, data_interface, interface_kwargs):
+    def test_deeplabcut_to_nwb(self, data_interface, interface_kwargs):
         nwbfile_path = self.savedir / f"{data_interface.__name__}.nwb"
 
         class TestConverter(NWBConverter):