+++ tests/test_ophys/test_tools_roiextractors.py
@@ -4,9 +4,12 @@ from tempfile import mkdtemp
 from pathlib import Path
 from datetime import datetime
 from copy import deepcopy
+from typing import Optional, List
+from types import MethodType
 
 import psutil
 import numpy as np
+from numpy.typing import ArrayLike
 from hdmf.data_utils import DataChunkIterator
 from hdmf.testing import TestCase
 from numpy.testing import assert_array_equal, assert_raises
@@ -279,6 +282,22 @@ class TestAddImageSegmentation(unittest.TestCase):
         self.assertEqual(image_segmentation.name, self.image_segmentation_name)
 
 
+def _generate_test_masks(num_rois: int, mask_type: str):  # Literal["pixel", "voxel"]
+    masks = list()
+    size = 3 if mask_type == "pixel" else 4
+    for idx in range(1, num_rois + 1):
+        masks.append(np.arange(idx, idx + size * idx, dtype=np.dtype("uint8")).reshape(-1, size))
+    return masks
+
+
+def _generate_casted_test_masks(num_rois: int, mask_type: str):  # Literal["pixel", "voxel"]
+    original_mask = _generate_test_masks(num_rois=num_rois, mask_type=mask_type)
+    casted_masks = list()
+    for per_roi_mask in original_mask:
+        casted_masks.append([tuple(x) for x in per_roi_mask])
+    return casted_masks
+
+
 class TestAddPlaneSegmentation(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
@@ -350,15 +369,9 @@ class TestAddPlaneSegmentation(unittest.TestCase):
 
         assert_array_equal(plane_segmentation_roi_centroid_data, expected_roi_centroid_data)
 
-        image_mask_iterator = plane_segmentation["image_mask"].data
-
-        data_chunks = np.zeros((self.num_rois, self.num_columns, self.num_rows))
-        for data_chunk in image_mask_iterator:
-            data_chunks[data_chunk.selection] = data_chunk.data
-
         # transpose to num_rois x image_width x image_height
         expected_image_masks = self.segmentation_extractor.get_roi_image_masks().T
-        assert_array_equal(data_chunks, expected_image_masks)
+        assert_array_equal(plane_segmentation["image_mask"], expected_image_masks)
 
     def test_do_not_include_roi_centroids(self):
         """Test that setting `include_roi_centroids=False` prevents the centroids from being calculated and added."""
@@ -434,6 +447,163 @@ class TestAddPlaneSegmentation(unittest.TestCase):
         accepted_roi_ids = list(np.logical_not(np.array(expected_rejected_roi_ids)).astype(int))
         plane_segmentation_accepted_roi_ids = plane_segmentation["Accepted"].data
         assert_array_equal(plane_segmentation_accepted_roi_ids, accepted_roi_ids)
+
+    def test_pixel_masks(self):
+        """Test the voxel mask option for writing a plane segementation table."""
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            num_rois=self.num_rois,
+            num_frames=self.num_frames,
+            num_rows=self.num_rows,
+            num_columns=self.num_columns,
+        )
+
+        def get_roi_pixel_masks(self, roi_ids: Optional[ArrayLike] = None) -> List[np.ndarray]:
+            roi_ids = roi_ids or range(self.get_num_rois())
+            pixel_masks = _generate_test_masks(num_rois=len(roi_ids), mask_type="pixel")
+            return pixel_masks
+
+        segmentation_extractor.get_roi_pixel_masks = MethodType(get_roi_pixel_masks, segmentation_extractor)
+
+        add_plane_segmentation(
+            segmentation_extractor=segmentation_extractor,
+            nwbfile=self.nwbfile,
+            metadata=self.metadata,
+            mask_type="pixel",
+        )
+
+        image_segmentation = self.nwbfile.processing["ophys"].get(self.image_segmentation_name)
+        plane_segmentations = image_segmentation.plane_segmentations
+
+        plane_segmentation = plane_segmentations[self.plane_segmentation_name]
+
+        true_pixel_masks = _generate_casted_test_masks(num_rois=self.num_rois, mask_type="pixel")
+        assert_array_equal(plane_segmentation["pixel_mask"], true_pixel_masks)
+
+    def test_voxel_masks(self):
+        """Test the voxel mask option for writing a plane segementation table."""
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            num_rois=self.num_rois,
+            num_frames=self.num_frames,
+            num_rows=self.num_rows,
+            num_columns=self.num_columns,
+        )
+
+        def get_roi_pixel_masks(self, roi_ids: Optional[ArrayLike] = None) -> List[np.ndarray]:
+            roi_ids = roi_ids or range(self.get_num_rois())
+            voxel_masks = _generate_test_masks(num_rois=len(roi_ids), mask_type="voxel")
+            return voxel_masks
+
+        segmentation_extractor.get_roi_pixel_masks = MethodType(get_roi_pixel_masks, segmentation_extractor)
+
+        add_plane_segmentation(
+            segmentation_extractor=segmentation_extractor,
+            nwbfile=self.nwbfile,
+            metadata=self.metadata,
+            mask_type="voxel",
+        )
+
+        image_segmentation = self.nwbfile.processing["ophys"].get(self.image_segmentation_name)
+        plane_segmentations = image_segmentation.plane_segmentations
+
+        plane_segmentation = plane_segmentations[self.plane_segmentation_name]
+
+        true_voxel_masks = _generate_casted_test_masks(num_rois=self.num_rois, mask_type="voxel")
+        assert_array_equal(plane_segmentation["voxel_mask"], true_voxel_masks)
+
+    def test_none_masks(self):
+        """Test the None mask_type option for writing a plane segementation table."""
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            num_rois=self.num_rois,
+            num_frames=self.num_frames,
+            num_rows=self.num_rows,
+            num_columns=self.num_columns,
+        )
+
+        add_plane_segmentation(
+            segmentation_extractor=segmentation_extractor, nwbfile=self.nwbfile, metadata=self.metadata, mask_type=None
+        )
+
+        image_segmentation = self.nwbfile.processing["ophys"].get(self.image_segmentation_name)
+        plane_segmentations = image_segmentation.plane_segmentations
+
+        plane_segmentation = plane_segmentations[self.plane_segmentation_name]
+        assert "image_mask" not in plane_segmentation
+        assert "pixel_mask" not in plane_segmentation
+        assert "voxel_mask" not in plane_segmentation
+
+    def test_pixel_masks_auto_switch(self):
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            num_rois=self.num_rois,
+            num_frames=self.num_frames,
+            num_rows=self.num_rows,
+            num_columns=self.num_columns,
+        )
+
+        def get_roi_pixel_masks(self, roi_ids: Optional[ArrayLike] = None) -> List[np.ndarray]:
+            roi_ids = roi_ids or range(self.get_num_rois())
+            pixel_masks = _generate_test_masks(num_rois=len(roi_ids), mask_type="pixel")
+            return pixel_masks
+
+        segmentation_extractor.get_roi_pixel_masks = MethodType(get_roi_pixel_masks, segmentation_extractor)
+
+        with self.assertWarnsRegex(
+            expected_warning=UserWarning,
+            expected_regex=(
+                "Specified mask_type='voxel', but ROIExtractors returned 3-dimensional masks. "
+                "Using mask_type='pixel' instead."
+            ),
+        ):
+            add_plane_segmentation(
+                segmentation_extractor=segmentation_extractor,
+                nwbfile=self.nwbfile,
+                metadata=self.metadata,
+                mask_type="voxel",
+            )
+
+        image_segmentation = self.nwbfile.processing["ophys"].get(self.image_segmentation_name)
+        plane_segmentations = image_segmentation.plane_segmentations
+
+        plane_segmentation = plane_segmentations[self.plane_segmentation_name]
+
+        true_voxel_masks = _generate_casted_test_masks(num_rois=self.num_rois, mask_type="pixel")
+        assert_array_equal(plane_segmentation["pixel_mask"], true_voxel_masks)
+
+    def test_voxel_masks_auto_switch(self):
+        segmentation_extractor = generate_dummy_segmentation_extractor(
+            num_rois=self.num_rois,
+            num_frames=self.num_frames,
+            num_rows=self.num_rows,
+            num_columns=self.num_columns,
+        )
+
+        def get_roi_pixel_masks(self, roi_ids: Optional[ArrayLike] = None) -> List[np.ndarray]:
+            roi_ids = roi_ids or range(self.get_num_rois())
+            voxel_masks = _generate_test_masks(num_rois=len(roi_ids), mask_type="voxel")
+            return voxel_masks
+
+        segmentation_extractor.get_roi_pixel_masks = MethodType(get_roi_pixel_masks, segmentation_extractor)
+
+        with self.assertWarnsRegex(
+            expected_warning=UserWarning,
+            expected_regex=(
+                "Specified mask_type='pixel', but ROIExtractors returned 4-dimensional masks. "
+                "Using mask_type='voxel' instead."
+            ),
+        ):
+            add_plane_segmentation(
+                segmentation_extractor=segmentation_extractor,
+                nwbfile=self.nwbfile,
+                metadata=self.metadata,
+                mask_type="pixel",
+            )
+
+        image_segmentation = self.nwbfile.processing["ophys"].get(self.image_segmentation_name)
+        plane_segmentations = image_segmentation.plane_segmentations
+
+        plane_segmentation = plane_segmentations[self.plane_segmentation_name]
+
+        true_voxel_masks = _generate_casted_test_masks(num_rois=self.num_rois, mask_type="voxel")
+        assert_array_equal(plane_segmentation["voxel_mask"], true_voxel_masks)
 
     def test_not_overwriting_plane_segmentation_if_same_name(self):
         """Test that adding a plane segmentation with the same name will not overwrite
@@ -1042,7 +1212,7 @@ class TestAddTwoPhotonSeries(TestCase):
         mock_imaging.get_num_frames.return_value = num_frames_to_overflow
 
         reg_expression = (
-            f"Memory error, full TwoPhotonSeries data is (.*?) GB are available! Please use iterator_type='v2'"
+            "Memory error, full TwoPhotonSeries data is (.*?) GB are available! Please use iterator_type='v2'"
         )
 
         with self.assertRaisesRegex(MemoryError, reg_expression):